fix bugs.

+++ python/paddle/fluid/tests/unittests/test_reshape_op.py
@@ -33,7 +33,8 @@ from op_test import OpTest
 # 
 #     def test_check_grad(self):
 #         self.check_grad(["X"], "Out")
-
+# 
+# 
 # class TestReshapeOpDimInfer1(OpTest):
 #     def setUp(self):
 #         self.op_type = "reshape"
@@ -56,7 +57,8 @@ class TestReshapeOp2(OpTest):
         self.op_type = "reshape"
         self.inputs = {
             "X": np.random.random(ori_shape).astype("float32"),
-            "Shape": np.array(new_shape)
+            "Shape": np.array(
+                new_shape, dtype="int64")
         }
         self.outputs = {"Out": self.inputs["X"].reshape(new_shape[0])}
 
@@ -66,6 +68,33 @@ class TestReshapeOp2(OpTest):
     def test_check_grad(self):
         self.check_grad(["X"], "Out")
 
+
+# class TestReshapeOpInplace(OpTest):
+#     def setUp(self):
+#         self.op_type = "reshape"
+#         self.inputs = {'X': np.random.random((10, 20)).astype("float32")}
+#         self.attrs = {'shape': [10 * 20], 'inplace': True}
+#         self.outputs = {'Out': self.inputs['X'].reshape(self.attrs['shape'])}
+# 
+#     def test_check_output(self):
+#         self.check_output()
+# 
+#     def test_check_grad(self):
+#         self.check_grad(["X"], "Out")
+# 
+# 
+# class TestReshapeOpDimInferInplace(OpTest):
+#     def setUp(self):
+#         self.op_type = "reshape"
+#         self.inputs = {'X': np.random.random((10, 20)).astype("float32")}
+#         self.attrs = {'shape': [4, -1, 5], 'inplace': True}
+#         self.outputs = {'Out': self.inputs['X'].reshape(self.attrs['shape'])}
+# 
+#     def test_check_output(self):
+#         self.check_output()
+# 
+#     def test_check_grad(self):
+#         self.check_grad(["X"], "Out")
 
 if __name__ == "__main__":
     unittest.main()
